refactor: centralize pseudonym creation
